Refine the importWrappedKey function implementation
This commit refines the importWrappedKey function to correctly
associate user security identifiers with imported keys.
When a key's authorization list specifies user_auth_type as
HWAuthType::PASSWORD, passwordSid is added to the authorization
list as USER_SECURE_ID. Similarly, if the user_auth_type is
HWAuthType::FINGERPRINT, biometricSid is added as USER_SECURE_ID.
However, if the user_auth_type contains HWAuthType::ANY, passwordSid
is added as USER_SECURE_ID.This ensures that imported keys are properly
bound to the user's security credentials, adhering to the specified
security requirements.

Additionally, In-Applet KeyDescription Parsing: The parsing of
KeyDescription is moved from the HAL to the applet. Because
KeyDescription is cryptographically bound to the encrypted wrapped key,
parsing it within the applet after successful decyption ensures greater
security and integrity. This change simplifies the
INS_FINISH_IMPORT_WRAPPED_KEY_CMD API, which now receives only six
parameters instead of eight, as key format and key parameters are no
longer sent from the HAL.

Bug: 434039170
FLAG: BUGFIX
Test: atest VtsAidlKeyMintTargetTest
      atest CtsKeystoreTestCases
Change-Id: I1929dc141dcf38caadadc87af210203d49302b98